+++ golang/kurtosis_core_rpc_api_bindings/api_container_service_grpc.pb.go
@@ -29,6 +29,9 @@ type ApiContainerServiceClient interface {
 	RegisterService(ctx context.Context, in *RegisterServiceArgs, opts ...grpc.CallOption) (*RegisterServiceResponse, error)
 	// Generates files inside the test volume on the filesystem for a container
 	GenerateFiles(ctx context.Context, in *GenerateFilesArgs, opts ...grpc.CallOption) (*GenerateFilesResponse, error)
+	// Tells the API container that the client has static files it would like the API container to know about
+	// The API container will respond with paths inside the enclave directory where the client should put its files
+	RegisterStaticFiles(ctx context.Context, in *RegisterStaticFilesArgs, opts ...grpc.CallOption) (*RegisterStaticFilesResponse, error)
 	// Copies static files that have been registered with the API container into the file namespace of the given service
 	LoadStaticFiles(ctx context.Context, in *LoadStaticFilesArgs, opts ...grpc.CallOption) (*LoadStaticFilesResponse, error)
 	// Starts a previously-registered service by creating a Docker container for it
@@ -94,6 +97,15 @@ func (c *apiContainerServiceClient) RegisterService(ctx context.Context, in *Reg
 func (c *apiContainerServiceClient) GenerateFiles(ctx context.Context, in *GenerateFilesArgs, opts ...grpc.CallOption) (*GenerateFilesResponse, error) {
 	out := new(GenerateFilesResponse)
 	err := c.cc.Invoke(ctx, "/api_container_api.ApiContainerService/GenerateFiles", in, out, opts...)
+	if err != nil {
+		return nil, err
+	}
+	return out, nil
+}
+
+func (c *apiContainerServiceClient) RegisterStaticFiles(ctx context.Context, in *RegisterStaticFilesArgs, opts ...grpc.CallOption) (*RegisterStaticFilesResponse, error) {
+	out := new(RegisterStaticFilesResponse)
+	err := c.cc.Invoke(ctx, "/api_container_api.ApiContainerService/RegisterStaticFiles", in, out, opts...)
 	if err != nil {
 		return nil, err
 	}
@@ -186,6 +198,9 @@ type ApiContainerServiceServer interface {
 	RegisterService(context.Context, *RegisterServiceArgs) (*RegisterServiceResponse, error)
 	// Generates files inside the test volume on the filesystem for a container
 	GenerateFiles(context.Context, *GenerateFilesArgs) (*GenerateFilesResponse, error)
+	// Tells the API container that the client has static files it would like the API container to know about
+	// The API container will respond with paths inside the enclave directory where the client should put its files
+	RegisterStaticFiles(context.Context, *RegisterStaticFilesArgs) (*RegisterStaticFilesResponse, error)
 	// Copies static files that have been registered with the API container into the file namespace of the given service
 	LoadStaticFiles(context.Context, *LoadStaticFilesArgs) (*LoadStaticFilesResponse, error)
 	// Starts a previously-registered service by creating a Docker container for it
@@ -223,6 +238,9 @@ func (UnimplementedApiContainerServiceServer) RegisterService(context.Context, *
 }
 func (UnimplementedApiContainerServiceServer) GenerateFiles(context.Context, *GenerateFilesArgs) (*GenerateFilesResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GenerateFiles not implemented")
+}
+func (UnimplementedApiContainerServiceServer) RegisterStaticFiles(context.Context, *RegisterStaticFilesArgs) (*RegisterStaticFilesResponse, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method RegisterStaticFiles not implemented")
 }
 func (UnimplementedApiContainerServiceServer) LoadStaticFiles(context.Context, *LoadStaticFilesArgs) (*LoadStaticFilesResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method LoadStaticFiles not implemented")
@@ -347,6 +365,24 @@ func _ApiContainerService_GenerateFiles_Handler(srv interface{}, ctx context.Con
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(ApiContainerServiceServer).GenerateFiles(ctx, req.(*GenerateFilesArgs))
+	}
+	return interceptor(ctx, in, info, handler)
+}
+
+func _ApiContainerService_RegisterStaticFiles_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(RegisterStaticFilesArgs)
+	if err := dec(in); err != nil {
+		return nil, err
+	}
+	if interceptor == nil {
+		return srv.(ApiContainerServiceServer).RegisterStaticFiles(ctx, in)
+	}
+	info := &grpc.UnaryServerInfo{
+		Server:     srv,
+		FullMethod: "/api_container_api.ApiContainerService/RegisterStaticFiles",
+	}
+	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
+		return srv.(ApiContainerServiceServer).RegisterStaticFiles(ctx, req.(*RegisterStaticFilesArgs))
 	}
 	return interceptor(ctx, in, info, handler)
 }
@@ -521,6 +557,10 @@ var ApiContainerService_ServiceDesc = grpc.ServiceDesc{
 		{
 			MethodName: "GenerateFiles",
 			Handler:    _ApiContainerService_GenerateFiles_Handler,
+		},
+		{
+			MethodName: "RegisterStaticFiles",
+			Handler:    _ApiContainerService_RegisterStaticFiles_Handler,
 		},
 		{
 			MethodName: "LoadStaticFiles",